feat:Issue-41:Rewrite daemon to improve architec..
Improves architechture by moving the scheduling to imperatives own service.
This makes the code more modular and easier to understand.

Breaking changes: None

Resolves: #41
