+++ monitoring-agent-daemon/src/services/mod.rs
@@ -1,4 +1,7 @@
 mod monitors;
 mod monitoringservice;
+mod schedulingservice;
 
 pub use monitoringservice::MonitoringService;
+pub use schedulingservice::SchedulingService;
+
